2.0 authentification ok et début creation service sur mes ressources

+++ src/characters/characters.controller.ts
@@ -1,14 +1,34 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  UseGuards,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CharactersService } from './characters.service';
 import { CreateCharacterDto } from './dto/create-character.dto';
 import { UpdateCharacterDto } from './dto/update-character.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { GetUser } from 'src/auth/get-user.decorator';
+import { User } from 'src/user/entities/user.entity';
 
 @Controller('characters')
 export class CharactersController {
   constructor(private readonly charactersService: CharactersService) {}
 
   @Post()
-  create(@Body() createCharacterDto: CreateCharacterDto) {
+  @UseGuards(AuthGuard())
+  create(
+    @Body() createCharacterDto: CreateCharacterDto,
+    @GetUser() user: User,
+  ) {
+    if (!user.admin) {
+      throw new UnauthorizedException('Droits admin nécéssaires');
+    }
     return this.charactersService.create(createCharacterDto);
   }
 
@@ -23,12 +43,24 @@ export class CharactersController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateCharacterDto: UpdateCharacterDto) {
+  @UseGuards(AuthGuard())
+  update(
+    @Param('id') id: string,
+    @Body() updateCharacterDto: UpdateCharacterDto,
+    @GetUser() user: User,
+  ) {
+    if (!user.admin) {
+      throw new UnauthorizedException('Droits admin nécessaires');
+    }
     return this.charactersService.update(+id, updateCharacterDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(AuthGuard())
+  remove(@Param('id') id: string, @GetUser() user: User) {
+    if (!user.admin) {
+      throw new UnauthorizedException('Droits admin nécessaires');
+    }
     return this.charactersService.remove(+id);
   }
 }
